[fix] remove old slots endpoint replaced by the new one to avoid sql injection
